finish reviewing activity and workflow code
